Fixed UI for new notifications
